Revert "Auth+splitComponents"
