+++ src/components/Footer.jsx
@@ -1,5 +1,4 @@
 import React from 'react'
-import '../styles.sass'
 import { Navbar } from 'react-bootstrap'
 
 class Footer extends React.Component {
@@ -8,7 +7,11 @@ class Footer extends React.Component {
       '© 2019 Exadel, Inc. | Privacy Policy | Terms & Conditions | Team4 Internship'
     return (
       <div className="fixed-bottom">
-            <footer className='footer'>{FOOTER_TEXT}</footer>
+        <Navbar bg="dark" variant="dark">
+          <Navbar.Brand>
+            <footer>{FOOTER_TEXT}</footer>
+          </Navbar.Brand>
+        </Navbar>
       </div>
     )
   }
