--- src/styles.sass
@@ -1,11 +1,0 @@
-$font-stack:    Helvetica, sans-serif
-$primary-color: #333
-
-footer
-  height: 50px
-  text-align: center
-  background: #343a40
-  color: white
-
-
-
